+++ rust/calculator/src/lib.rs
@@ -1,29 +1,29 @@
-fn primes(num: usize) -> Vec<usize> {
+fn primes(num: u64) -> Vec<u64> {
     if num < 2 {
         return vec![];
     }
-    let mut primes: Vec<bool> = vec![true; num + 1];
+    let mut primes: Vec<bool> = vec![true; (num + 1) as usize];
     primes[0] = false;
     primes[1] = false;
 
-    let limit: usize = (num as f64).sqrt() as usize;
+    let limit: u64 = (num as f64).sqrt() as u64;
     for i in 2..=limit {
-        if primes[i] {
-            for j in (i * i..=num).step_by(i) {
-                primes[j] = false;
+        if primes[i as usize] {
+            for j in (i * i..=num).step_by(i as usize) {
+                primes[j as usize] = false;
             }
         }
     }
     primes
         .iter()
         .enumerate()
-        .filter_map(|(i, &is_prime)| if is_prime { Some(i) } else { None })
+        .filter_map(|(i, &is_prime)| if is_prime { Some(i as u64) } else { None })
         .collect()
 }
 
 #[unsafe(no_mangle)]
-pub extern "C" fn calc_primes(num: usize, out_len: *mut usize) -> *mut usize {
-    let primes: Vec<usize> = primes(num);
+pub extern "C" fn calc_primes(num: u64, out_len: *mut usize) -> *mut u64 {
+    let primes: Vec<u64> = primes(num);
     let len: usize = primes.len();
     // out_lenがNULLでない場合に要素数を書き込む
     unsafe {
@@ -31,15 +31,15 @@ pub extern "C" fn calc_primes(num: usize, out_len: *mut usize) -> *mut usize {
             *out_len = len;
         }
     }
-    let mut boxed_primes: Box<[usize]> = primes.into_boxed_slice();
-    let ptr: *mut usize = boxed_primes.as_mut_ptr();
+    let mut boxed_primes: Box<[u64]> = primes.into_boxed_slice();
+    let ptr: *mut u64 = boxed_primes.as_mut_ptr();
     // 所有権を放棄
     std::mem::forget(boxed_primes);
     ptr
 }
 
 #[unsafe(no_mangle)]
-pub extern "C" fn free_primes_array(ptr: *mut usize, len: usize) {
+pub extern "C" fn free_primes_array(ptr: *mut u64, len: usize) {
     if ptr.is_null() {
         return;
     }
